Drop consequent accesses to the same memory locations within a basic block.
For example, "i++" is translated into the sequence of load/add/store
instructions. Load and store access the same memory location with the same
access size, so we may skip the load instruction during the instrumentation.

If the current instruction is a store, we check that none of the previous
stores and/or loads in the basic block access the same memory location (we
exploit LLVM alias analysis to tell that). If not, the previous access is
removed from the "to be instrumented" set.
If the current instruction is a load, we check that none of the previous loads
access the same memory.
The new access is added to the TBI set anyway.

If a load occurs after a store, we instrument both.